requirements.txt
openai-whisper
setuptools-rust
speechrecognition
torch
torchaudio
numpy
pyaudio
soundfile
librosa
pvporcupine
pyaudio
pyttsx3